pyproject.toml
[project]
name = "dvr-network"
version = "1.0.0"
description = "This project is a Virtual DVR to connect in IP camera, read the stream video, chunck video, save and server in API REST."
authors = [
    {name = "wsalisonxp@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.13"
dependencies = [
    "uvloop (>=0.22.1,<0.23.0)",
    "tortoise-orm (>=1.1.7,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "aioboto3 (>=15.5.0,<16.0.0)"
]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false